Add support for keystoneauth Session objects
When using Craton with Keystone for identity and access management, we
need to tell keystoneclient and keystoneauth Session's what service
we're trying to talk to.

This also means that when using cratonclient's Session, we need to only
set up Craton's auth headers when username, token, and project_id are
set.

Change-Id: I69a6a742832c3571523d30a04be1cc4bd3da1e7b