WIP: CLI and client support for get/set of resource variables
Functionality is supported both	for the	Python client:

host = craton.hosts.get(item_id=8)
host.variables.update(x='foo', y={'a': 47, 'b': True}, z='baz')

As well	as for the CLI:

$ craton region-get-vars 1
$ craton cell-set-vars 8 x=true y=47 z=foo
$ craton host-set-vars 10 < baz.json
$ craton host-get-vars --format=json 10 | jq -C

etc.

Current resource setters are region-get-vars, region-set-vars,
cell-get-vars, cell-set-vars, host-get-vars, host-set-vars. Code is
implemented to minimize the overhead of doing this for additional
resources.

It would be more ergonomic if the CLI supported referencing a
resource, like a host, by an identifying name, not just an ID, but
that can be done in a separate branch.

CURRENTLY NO UNIT TESTING!!! This needs to be fixed before this work
can be merged in.

Change-Id: Id30188937518d7103d6f943cf1d038b039dc30cc