Remove sort arguments from host-list and cell-list
This commit updates the cratonclient commands host-list and cell-list to
remove the optional sort arguments --sort-dir and --sort-key because
they cause the client requests to fail.

craton does not support sorting. The sort arguments in cratonclient
appeared to work on first glance because craton was ignoring unexpected
parameters. The craton commit 07178e034712d9b6b1853dc9da7d79ee668c7290
now causes craton to return a 400 error if anything in the request is
not allowed.

host-list and cell-list were also failing because by default
sort_dir=asc was being added to the query, e.g.

/v1/hosts?sort_dir=asc

Change-Id: I06e6fc1b618cd4c01431ad8dd3a1ec360d5e3b02
Related-bug: 1664513