Relax requirement of region-id for hosts and cells
When listing hosts and cells, there is no need to require the user to
provide a region-id. The API previously required region IDs to be
present in the query string to list hosts and cells, but relaxed this
constraint. The craton CLI encoded this constraint in its parameters but
now needs to relax this as well.

Change-Id: I973176a150c5d10de2144501dcc39cd6cfacbf9a
Closes-bug: #1659428