Add table formatter sort key
There are many attributes, such as parent_id, where craton uses None to
represent the absence of a value. This changes allows PrettyTable to
sort data in Python 3 where it is either of a type, e.g. int, or it is
None.

Python 3 fails when attempting to sort values of different types. In
Python 2.7 sorting [1, None, 2] gives [None, 1, 2], in Python 3 a
TypeError is raised. This change adds a sort key that enables a
comparison based on whether or not the value is None before attempting
to compare the values themselves.

Change-Id: I8219eadab052e103d2089ebb69ac84f1df795f02
Closes-bug: 1668221