Add basic shell documentation for cratonclient
This adds a directory and some basic templates for creating shell
command documentation for python-cratonclient. This creates the shell
sub-directory to allow for individual pages which document each
sub-command in detail.

In the future, that documentation should be further scoped by the
version of the API that it happens to be relevant to. At the moment that
is not a concern so the flatter structure should work fine.

Change-Id: I07bb776c7713283283c86a2907e1c0ba1c763b79