Update link to documentation and bug tracker
ReadTheDocs provides TLS for all subdomains and with our updated
infrastructure jobs, we are updating the documentation here on every
push to master.

Further, we split out bug tracking for craton's API and it's client code
a while ago, this just points people at the right place to report those
bugs now.

Change-Id: Id946bba0cda9e4aa81d6df9088e3b18afaca8ee9